update L2 projection (#300)
* L2 projection: change internal DofHandler to MixedDofHandler

This makes the L2 projection usable for grids with several celltypes. It 
also adds the possibilty to give a 2nd CellValues object for computing 
the mass matrix. This is useful when the regular CellValues don't have 
enough integration points for correctly integrating N*N.

* enable project_to_nodes when using subset of mesh or SymmetricTensors

project_to_nodes will store NaNs in all nodes that do not belong to the 
subset. The project_to_nodes option is added to the project function for 
SymmetricTensors and by default true (like to regular Tensors).

* bugfix L2-projection of SymmetricTensors

Use the independent elements of the Symmetric Tensor instead of the 
first 1:M elements in linear indexing.

* additional test for L2-projection

tests the L2-projection on a mixed grid and for a field that lives on a 
subset of the domain (and that is a SymmetricTensor field).

* add type for node2dof_map

* using single project function

use the same project function for SymmetricTensor and Tensor

* introduce new L2Projector constructor

constructor based on quadrature rule and interpolation

* comment type for M_cholesky back in 

Why was this type commented out? It seems to work.

* comment in type for M_cholesky

No, it doesn't work.

* fix tensor construction in project()

* fix 1.0 compatibilty

* introduce __close! for both DofHandlers

* fix bug