Evaluation of FE-approximated function in any point (#377)
Evaluation of FE-approximated function in any point

This adds a new feature allowing for evaluation of the FE-approximated
function in arbitrary points in the domain. The procedure has two
steps:
  1. Creating a `PointEvalHandler` using the grid and the points of
     interest. This will lookup (i) the corresponding cell, and (ii)
     the local coordinate corresponding to every point.
  2. Any approximated function can now be evaluated at these points
     with the `get_points_values` function, which takes (i) the
     `PointEvalHandler`, a `DoFHandler`, and a vector of dof-values,
     and return a vector with the approximation evaluated at the
     points of the `PointEvalHandler`.

This work was initiated by Jack Muir (at-jbmuir) (#316), before
Kim Louisa Auth (at-kimauth) and Maximilian Köhler (at-koehlerson)
took over and did the lion's share of the work. The final API-changes
and touch-ups was done by Kim Louisa Auth and Fredrik Ekre (at-fredrikekre).


Co-authored-by: Jack B. Muir <jmuir@caltech.edu>
Co-authored-by: Kim Louisa Auth <auth@chalmers.se>
Co-authored-by: Maximilian Köhler <maximilian.koehler@ruhr-uni-bochum.de>
Co-authored-by: Fredrik Ekre <ekrefredrik@gmail.com>