Add support for affine constraints on dofs (#401)
This adds the AffineConstraint constraint type that can be added to the
constraint handler. An affine constraint of the form
    u_i = \sum_j (c_j * u_j) + b,
where u_i are dofs, c_j scaling factors and b a inhomogeneity is
constructed and added to the constraint handler as
    ac = AffineConstraint(i, [u_j => c_j, ...], b)
    add!(ch, ac)

Sparsity patterns must then be constructed with the constraint handler
e.g. create_sparsity_pattern(dh, ch), and apply! must be used after
solving the system (apply!(u, ch), where u is the solution).