PBC: Change definition of mirror/image mapping (#501)
This patch changes the meaning of the mapping in collect_periodic_faces
and the rotation matrix in PeriodicDirichlet. Previously these
transformations acted on the mirror to transform into the corresponding
entities on the image, but now it is the opposite. This is to match how
this is usually defined in literature.