Improvements to coloring
This patch generalizes the coloring algorithms to accept any iterable
for the subset of cells to color. In particular this means that there is
no need to materialize the cellset when coloring all cells in the grid.
In addition this means that, when coloring all the elements, the result
should be deterministic instead of relying on the iteration order of a
Set{Int}.

Fixes #502.