Local condensation of constraints
This patch adds functionality for applying constraints locally on the
element matrix/vector before assembling them into the global
matrix/vector. When doing local condensation there is no need to apply
constraints on the global system after the assembly. The main motivation
for this functionality is when using "write-only" global arrays (for
example some array type in an external library such as HYPRE) where it
is expensive or impractical to modify the array after assembling.

Local constraint condensation can be done with apply_local! or
apply_assemble!:

    apply_local!(local_matrix, local_vector, global_dofs, ch)

apply_local! can be used when all constraints are local to the element
(e.g. Dirichlet boundary constraints) since the constraints do not
couple to dofs outside the dofs of the element itself.

    apply_assemble!(assembler, global_dofs, local_matrix, local_vector)

apply_assemble! condensates constraints and assembles the result into
the global system, i.e. performs apply_local! + assemble!. The advantage
with apply_assemble! is that it handles general constraints since it has
access to the global matrix/vector and can write to entries outside the
local matrix when necessary.