Remove `create_sparsity_pattern(::MixedDofHandler)` (#538)
This patch removes the special method for creating a sparsity pattern of
a `MixedDofHandler` and slightly generalizes the one for `DofHandler`.
The removed method was almost identical to the one for `DofHandler`
except that it allocated a new global dof vector with the correct size
for every element. The only necessary generalization is thus to call
`resize!` on the global dof vector for every element. This should result
in fewer allocations when creating a sparsity pattern for a
`MixedDofHandler`.