Specifying coupling between fields/components in `create_sparsity_pattern`
This patch implements a new keyword argument `coupling` to
`create_sparsity_pattern` where the coupling between fields or
components in the DofHandler can be specified. `coupling` should be a
matrix of booleans, where rows are test functions and columns the
unknowns. Consider for example a system with (u, p) as unknowns, and (v,
q) as the corresponding test functions. If there is no coupling between
e.g. p and q the coupling matrix would be
        u      p
        -------------
    v  | true  true
    q  | true  false

i.e. `coupling = [true true; true false]`. The coupling can also be
specified component wise (if one of the fields have multiple
components). For example, if (u, v) have two components, the equivalent
coupling specified by components would be
         u1     u2    p
        ------------------
    v1 | true  true  true
    v2 | true  true  true
    q  | true  true  false

The resulting sparsity pattern will not have entries for components that
do not couple. By default full coupling is assumed (just like
before).

The new functionality is used in the stokes-flow.jl example, where there
are no coupling between test- and trial functions for the pressure.