Add internal function `fillzero!` for filling stored entries with 0
This patch adds a new internal function `fillzero!(A)` which fills
(stored) entries of `A` with 0. This is part of breaking down the
assembly procedure into smaller pieces such that each piece can be
overriden more easily by other array types, similar to #548.
`fillzero!(A)` falls back to `fill!(A, 0)`.