Rework IJVAssembler to a full-fledged assembler
This patch reworks the IJVAssembler to function more like a standard
assembler. In particular, it now supports assembling local vectors, and
supports assembling with local application of constraints.

Note also that `finish_assemble` now return both the matrix and the
vector. The vector will be empty if the assembler is used for the matrix
entries only.