Fix constraint condensation for `Symmetric` matrices.
This patch fixes an issue where `f` would be incorrectly modified in
`apply!(K, f, ch)` when `K::Symmetric` and when using inhomogeneous
constraints. Specifically this patch make sure that entries below the
diagonal of the matrix backing the `Symmetric` aren't used, and that the
corresponding entries from above the diagonal are used in their place.