Store field names in MixedDofHandler (#646)
This patch adds a field `field_names::Vector{Symbol}` to
`MixedDofHandler` that caches the computation of global field names,
i.e. `getfieldnames(dh)`. This mirrors how fields are stored in the
DofHandler. The usage of the two similar functions,
`getfieldinterpolations` and `getfielddims`, should be investigated and
possibly it might make sense to store these directly in the
(Mixed)DofHandler too.