Create a `BitSet` of `FieldHandler.cellset` in loops (#656)
This patch creates a `BitSet` of `FieldHandler.cellset` in loops, in
particular in `close!(::DofHandler)` and
`create_sparsity_pattern(::DofHandler)`. Since `BitSet`s are sorted this
ensures that these loops are done in ascending cell order, which gives a
performance boost due to better memory locality.

This is an even smaller change than #654 (and #625) which should be
completely non-breaking since the type of `FieldHandler.cellset` is not
changed. Larger refactoring, such as using `BitSet` or `OrderedSet` will
be done in the `FieldHandler/`SubDofHandler` rework.