Add mapping from cell id to field handler id
This patch adds a mapping in the `MixedDofHandler` struct from cell id
to `FieldHandler` id. This makes it possible to store data in the
`FieldHandler` struct, and still fetch it given only the dof handler and
the cell id. For example, the number of dofs per cell is constant within
each field handler, but currently it is stored for every cell in the dof
handler. In fact, this vector is replaced in this PR, so memory wise the
mapping is for free, but strictly more useful.

The extra indirection makes `ndofs_per_cell` approximately twice as
expensive. The following benchmark, on a 1000x1000 grid, goes from 280μs
to 583μs but ¯\_(ツ)_/¯.

```
function cound_dofs(dh)
    c = 0
    for i in 1:ncells
        c += ndofs_per_cell(dh, i)
    end
    return c
end
```