Use `BitSet` in `apply_analytical!` and `reshape_to_nodes`
This patch uses `BitSet` in `apply_analytical!` and `reshape_to_nodes`
for `MixedDofHandler`. The benefit here is twofold: computing the
intersection is much faster (basically just bitwise `&`) and the
subsequent looping over the cells are done in ascending cell order.

This closes the performance gap between `MixedDofHandler` and
`DofHandler` in benchmarks from #629 of `apply_analytical!`,
`reshape_to_nodes`, and `vtk_point_data`. For example, here is the
benchmark results for `apply_analytical!`:
```
387.853 ms (72 allocations: 34.50 MiB)  # MixedDofHandler master
 55.262 ms (38 allocations: 553.45 KiB) # MixedDofHandler patch
 41.861 ms (14 allocations: 2.27 KiB)   # DofHandler master/patch
```