Change interpolation storage from vector-of-structs to struct-of-vectors
This changes the storage of field information in `FieldHandler` from a
`Vector{Field{Symbol,Int,Interpolation}}` to three vectors
`Vector{Symbol}`, `Vector{Int}`, `Vector{Interpolation}`. The `Field`
struct will (likely) be removed in the `FieldHandler` -> `SubDofHandler`
rework. Based on the (limited) usage in the code base this also seem
more natural, and this is also how interpolations were stored in the old
`DofHandler`. For now the old field is kept to help with compatibility.

Closes #669.