Rework code for `add!(::ConstraintHandler, ::Dirichlet)` (#672)
This patch cleans up the code for adding `Dirichlet` constraints to the
`ConstraintHandler` a bit after the DofHandler merge. In particular, it
deprecates the method
```
add!(::ConstraintHandler, ::FieldHandler, ::Dirichlet)
```
in favor of just
```
add!(::ConstraintHandler, ::Dirichlet)
```
which was added in #427. There is no need to be able to specify the
constrained set by both the set passed to `Dirichlet` and by the set
given to the `FieldHandler`. Looking at the tests modified in this
patch, it seems this was never the intention anyway. This patch also
removes the copy of the set in the `Dirichlet` constructor in favor of
explicit partitioning of the set over the `FieldHandler`s.