Update VTK export to use CellValues for evalutation
This patch deprecates `reshape_to_nodes` in favor of the new
`evaluate_at_grid_nodes` function which more or less does the same
thing. `evaluate_at_grid_nodes` uses the interpolation of the field to
evaluate the approximated function in the cell coordinates (by using
`reference_coordinates` of the cell as quadrature points). This is much
simpler, and more correct, compared to figuring out which Dofs live on
which node etc. In particular, this now support non-standard
interpolations which may not even have dofs in the nodes.

Fixes #167.