Merge `(Cell|Face|Point)(Scalar|Vector)Values` into `(Cell|Face|Point)Values`
This patch merges:
 - `CellScalarValues` and `CellVectorValues` into `CellValues`,
 - `FaceScalarValues` and `FaceVectorValues` into `FaceValues`,
 - `PointScalarValues` and `PointVectorValues` into `PointValues`.

This is possible now that interpolations are vectorized instead (#694)
and, for example, `CellVectorValues(::Interpolation)` can now be spelled
as simply `CellValues(::VectorInterpolation)`.

The new structs have new parametrization to be more general wrt the
storage types to enable embedded elements (#651). They are also
parametrized wrt the function interpolation, to enable dispatching on
this in e.g. `reinit!`.

Instead of simply deprecating the old API (with a warning) this  patch
implements hard errors for the old constructors with a clear message on
how to upgrade. The reason for this is that i) it is possible to tank
performance with the new parametrization, ii) for any non-trivial use
one would run into errors anyway (e.g. `f(::CellVectorValues)` would give
a non-descriptive `MethodError`).

Closes #682.