Use the new reference shapes for `Interpolation`s and `QuadratureRule`
This patch replaces `RefCube{1, 2, 3}` with `RefLine`,
`RefQuadrilateral`, and `RefHexahedron`; and `RefTetrahedron{2}` with
`RefTriangle` in the definitions of interpolations and quadrature.

Since the reference dimension is now encoded in the shape, this patch
also changes the parameter order of interpolations, to move the dim
parameter to the end, and always set it to `Nothing`. It can't be
removed completely right away since that would give cryptic error
messages.

Interpolations are thus now constructed with just the reference shape
and the order, e.g. `Lagrange{RefTriangle, 2}()` instead of `Lagrange{2,
RefTetrahedron, 2}()`.

The same is done for `QuadratureRule`, for constructing cell quadrature
rules. A face quadrature rule is still written as `QuadratureRule{D-1,
shape}(...)`, where `shape` is the reference shape of the *cell*, i.e.
face quadrature for the hexahedron is constructed as `QuadratureRule{2,
RefHexahedron}()`. The new reference shapes make this non-ambiguous
since we can now directly that a face quadrature is constructed.