`QuadratureRule`: remove `dim` parameter and introduce `FaceQuadratureRule` (#716)
This patch removes the `dim` parameter from `QuadratureRule`
constructors, since it is now implicit from the reference shape.

Unlike for interpolations (#711) the `dim` parameter for
`QuadratureRule` was not completely redundant since it was used to
construct quadrature rules for face integration. This patch therefore
introduces `FaceQuadratureRule` as a replacement:
`FaceQuadratureRule{refshape}` instead of `QuadratureRule{dim - 1,
refshape}`. The new syntax generalizes better to elements such as wedge
or pyramid which doesn't have a unique `dim - 1` reference shape.

With the new syntax it is also possible to materialize the rule for each
face already in the `FaceQuadratureRule` constructor, instead of doing
it in the `FaceValues` constructor.