Adopt proper codim/dim terminology for entities (#789, #914)
This patch changes the naming and meaning of various entities in order
to be independent of the current dimension of the problem/grid/element
type.

After this patch, the following terminology is used to name entities
based on their dimension:

 - `vertex`: a zero-dimensional entity (point, node)
 - `edge`: a one-dimensional entity between two vertices
 - `face`: a two-dimensional entity between edges
 - `volume`: a three-dimensional entity between faces

and the following terminology is used to name entities based on their
codimension:
 - `cell`: codimension zero (same meaning as before this patch)
 - `facet`: codimension one

Note that the codimension is defined relative the *reference dimension*
of the cell. For example, in a problem in three spatial dimensions with
both three-dimensional hexahedrons and two-dimensional quadrilaterals,
the facets of the hexahedron are faces, and the facets of the
quadrilateral are edges.

These changes simplifies internal handling of degree-of-freedom
distribution, and simplifies handling of boundary conditions (which are
now consistently dealing with *facets* instead of *faces*, *edges*, and
*vertices for three-, two-, and one-dimensional problems respectively.

User-facing changes:

 - Face-sets in the `Grid` type are deprecated in favor of facet-sets:
   (`getfaceset` is replaced by `getfacetset`, `addfaceset!` is replaced
   by `addfacetset!`).
 - Facet-sets should be used instead of face-sets when defining boundary
   conditions (such as `Dirichlet`).
 - `FaceValues` have been renamed to `FacetValues`.
 - `FaceQuadratureRule` have been renamed to `FacetQuadratureRule`.
 - `nfaces` should be replaced with `nfacets` (when looping over the
   facets of an element to integrate Neumann boundary conditions, for
   example).

Fixes #899, fixes #901. Closes #789, closes #914.

Co-authored-by: Elias <elias.isak.borjesson@gmail.com>
Co-authored-by: Knut Andreas Meyer <knutam@gmail.com>