Abstraction layer around VTK export (#692)
This patch introduces a Ferrite-specific abstraction layer on top of
WriteVTK.jl instead of re-exporting and extending methods.

VTK files are created with the `VTKFile` constructor (instead of
`WriteVTK.vtk_grid`) and closed with `Base.close`. To add data to the
file the following functions, which replaces methods of
`WriteVTK.vtk_point_data` and `WriteVTK.vtk_cell_data`, should be used:

 - `write_solution(::VTKFile, ::DofHandler, ::Vector)`: write a solution
   vector corresponding to DoFs in the DofHandler.
 - `write_projection(::VTKFile, ::L2Projector, ::Vector)`: write a
   projected vector corresponding the DoFs in the DofHandler of the
   projector.
 - `write_cell_data(::VTKFile, ::Vector)`: write data per cell.
 - `write_node_data(::VTKFile, ::Vector)`: write data per node.
 - `Ferrite.write_cellset`, `Ferrite.write_nodeset`: visualize cellsets
   and nodesets.
 - `Ferrite.write_constraints`: visualize constrained degrees of
   freedom.
 - `Ferrite.write_cell_colors`: visualize the result of grid coloring.

To collect multiple time steps the `VTKFileCollection` struct is used
(instead of `WriteVTK.paraview_collection`). VTK files are added to the
collection with `addstep!`.

Fixes #278, fixes #686. Closes #691.