Deprecate `onboundary`, remove `boundary_matrix` property from `Grid` (#924)
This patch deprecates the `onboundary` function which checks whether an
element has any facet alone an external boundary. This functionality
relied on the `boundary_matrix` property in `Grid`, but this was only
populated correctly by the builtin grid generators (and not e.g. for
meshes parsed from other formats).

Checking `onboundary` was just a shortcut to avoid some extra work in
the facetset containment check which you always had to do anyways, but
since this is just a tuple hash it isn't expensive anyway.

In addition, nowadays there are utilities such as `FacetIterator` which
let's you explicitly loop over the facet sets you want directly.