Replace `shape_value` etc for interpolations with `reference_shape_value`
This patch is a follow up to #721 and replaces `shape_value`,
`shape_gradient` etc for interpolations, which return evaluations in the
reference domain, with new functions `reference_shape_value`,
`reference_shape_gradient`, etc. The new functions are public but not
exported.

Separating these functions make it possible to support
`reference_shape_value` etc for `CellValues` too in the future.