Deprecate assemble! with local vector before local matrix
This patch deprecates the signature
`assemble!(assembler, dofs, local_vector, local_matrix)` in favor of
`assemble!(assembler, dofs, local_matrix, local_vector)`. The former
signature was not used in many places, not supported by e.g. the block
matrix assembler, and it is confusing that there are two options. See
also #707 for a similar change to `start_assemble`.

In addition, if the assembler is used just for matrix assembly, this
patch removes an unnecessary allocation of an empty vector for every
call to `assemble!`.