Clarify construction of Jacobian in Navier-Stokes tutorial
In the incompressible Navier-Stokes tutorial, there are two things wish
I noticed that might be improved:

P1. The construction of the Jacobian is a bit hand-wavy in the sense that there
is a claim that the Jacobian and stiffness matrix share the same
sparsity pattern.

P2. It is not necessarily clear what motivates the `navierstokes_jac_element!`
function.

To address these points, I have added a small derivation for the analytical form
of the Jacobian. In particular, I note that the nonlinear
advection term must be handled explicitly when constructing the
Jacobian. The functions `navierstokes_jac!` and
`navierstokes_jac_element!` already implement what I describe in the
derivation, but to make things explicit, I believe the derivation is
worth including.

The derivation is currently in the section
`Commented Implementation > Solution of the semi-discretized system via DifferentialEquations.j` and may be more appropriate in the theory section.